+++ gulp/tasks/watch.js
@@ -1,6 +1,6 @@
 'use strict';
 var gulp = require('gulp');
-var config = require('./config');
+var config = require('../config');
 
 gulp.task('watch', function() {
   gulp.watch([config.Path.CSS_SOURCES], ['sass']);
