+++ gulpfile.js
@@ -1,15 +1,15 @@
+var fs = require('fs');
 var gulp = require('gulp');
-var wrench = require('wrench');
 
 /**
  * This will load all js or coffee files in the gulp directory
  * in order to load all gulp tasks
  */
-wrench.readdirSyncRecursive('./gulp').filter(function(file) {
+fs.readdirSync('./gulp/tasks').filter(function(file) {
   return (/\.(js|coffee)$/i).test(file);
 }).map(function(file) {
-  require('./gulp/' + file);
+  require('./gulp/tasks/' + file);
 });
 
 gulp.task('build', ['build_js', 'sass']);
-gulp.task('default', ['build_js', 'sass', 'watch']);
+gulp.task('default', ['build', 'watch']);
