+++ package.json
@@ -13,7 +13,6 @@
     "gulp-plumber": "^1.0.0",
     "gulp-sass": "^2.1.1",
     "gulp-uglify": "^1.2.0",
-    "run-sequence": "^1.0.2",
-    "wrench": "^1.5.8"
+    "run-sequence": "^1.0.2"
   }
 }
